Added respeaker code
This code adds visualizations that display input volume level
using the leds on top of the SeeedStudio Respeaker 4 mic
Raspberry Pi hat. This device has 12 leds on top of it which
are used to display the current volume level, the threshold
level for Activity Detection, and whether or not the assistant
is currently recording.